move cypher functions to own file, add 'config' kwarg to pyingest
config arg will replace yaml_string in future releases. it is more descriptive and inclusive of the newly allowed formats for yaml input
