+++ neo4j_runway/ingestion/pyingest.py
@@ -4,7 +4,8 @@ This is a modified PyIngest file for Neo4j Runway. It currently only supports Pa
 
 import os
 import datetime
-from typing import Optional
+from typing import Optional, Union
+import warnings
 
 from neo4j import GraphDatabase
 import numpy as np
@@ -73,7 +74,6 @@ class LocalServer(object):
             skip = params["skip_file"] if "skip_file" in params else False
             if skip:
                 return
-            print("curdir: ", os.listdir())
             with open(params["url"]) as openfile:
                 # Grab the header from the file and pass that to pandas.  This allow the header
                 # to be applied even if we are skipping lines of the file
@@ -129,7 +129,11 @@ def load_config(configuration):
     config = yaml.safe_load(configuration)
 
 
-def PyIngest(yaml_string: str, dataframe: Optional[pd.DataFrame] = None) -> None:
+def PyIngest(
+    yaml_string: Union[str, None] = None,
+    dataframe: Optional[pd.DataFrame] = None,
+    **kwargs
+) -> None:
     """
     Function to ingest data according to a configuration YAML string.
     This is a modified lite version of the original PyIngest.
@@ -139,11 +143,22 @@ def PyIngest(yaml_string: str, dataframe: Optional[pd.DataFrame] = None) -> None
     yaml_string : str
         A string representation of the YAML file that is generated by the IngestionGenerator class.
         May also be a filepath to a YAML file.
+        (Will be depreciated. Use the config kwarg instead.)
     dataframe : Optional[pd.DataFrame], optional
         The data to ingest in Pandas DataFrame format.
         If None, then will search for CSVs according to the urls in YAML config, by default None
+    config : str
+        A string representation of the YAML file that is generated by the IngestionGenerator class.
+        May also be a filepath to a YAML file.
+
     """
-    load_config(get_yaml(yaml_string))
+    if "config" in kwargs:
+        load_config(get_yaml(kwargs["config"]))
+    else:
+        load_config(get_yaml(yaml_string))
+        warnings.warn(
+            "the yaml_string parameter will be depreciated in future releases. Please use the 'config' kwarg to identify the YAML file instead."
+        )
     server = LocalServer()
     server.pre_ingest()
     file_list = config["files"]
