+++ neo4j_runway/tests/test_code_generator/test_ingest_generation.py
@@ -1,7 +1,8 @@
 import unittest
 
+from ...ingestion.cypher import *
+from ...ingestion.generate_ingest import IngestionGenerator
 from ...objects import Node, Relationship, Property, DataModel
-from ...ingestion.generate_ingest import *
 from ...tests.resources.ingestion_generation_answers import *
