+++ neo4j_runway/tests/test_code_generator/test_pyingest_post_ingest_input.py
@@ -1,9 +1,8 @@
+from typing import List
 import unittest
 
 from ...objects import Node, Relationship, Property, DataModel
-from ...ingestion.generate_ingest import *
-
-# from ...tests.resources.ingestion_generation_answers import *
+from ...ingestion.generate_ingest import IngestionGenerator
 
 
 nodes = [
